apps/server/python/requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
einops==0.8.1
filelock==3.20.1
fsspec==2025.12.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.0
packaging==25.0
pillow==12.0.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
sympy==1.14.0
timm==1.0.22
tokenizers==0.22.1
torch==2.9.1+cpu
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
urllib3==2.6.2
